\name{FABDataHandler-package}
\alias{FABDataHandler-package}
\alias{FABDataHandler}
\docType{package}
\title{
\packageTitle{FABDataHandler}
}
\description{
\packageDescription{FABDataHandler}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FABDataHandler}
\packageIndices{FABDataHandler}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{FABDataHandler}

Maintainer: \packageMaintainer{FABDataHandler}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
